requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.25.2
torch
torchvision
pandas
requests
seaborn